+++ libs/gpu_md5/src/cuda/gpu_string_generator.cuh
@@ -5,11 +5,11 @@
 CDECL
 struct GpuStringGenerator{
       char current[24];
-      int initialSequenceLength;
-      int currentSequenceLength;
+      uint8_t initialSequenceLength;
+      uint8_t currentSequenceLength;
       bool current_used;
 };
-__device__ GpuStringGenerator new_generator(int initialSequenceLength);
+__device__ GpuStringGenerator new_generator(uint8_t initialSequenceLength);
 __device__ void assign_address(GpuStringGenerator *gen, size_t address);
 __device__ void generate_chunk(GpuStringGenerator *gen, size_t size,
                                char *data);
