+++ libs/gpu_md5/src/cuda/md5transform.cuh
@@ -5,7 +5,7 @@
 CDECL
 struct md5_transform_request {
   uint8_t *data;
-  uint32_t *sizes;
+  uint8_t *sizes;
   char *result;
   uint32_t *offsets;
   size_t num_of_strings;
@@ -13,7 +13,7 @@ struct md5_transform_request {
 };
 
 
-struct md5_transform_request new_request(uint8_t *data, uint32_t *sizes, size_t num_of_strings);
+struct md5_transform_request new_request(uint8_t *data, uint8_t *sizes, size_t num_of_strings);
 void free_request(struct md5_transform_request *req);
 void md5_gpu_transform(struct md5_transform_request request, int maxthreads);
 END
